stop infinite loop
