+++ components/site-header.tsx
@@ -18,9 +18,10 @@ export async function SiteHeader() {
     data: { user },
   } = await supabase.auth.getUser();
 
-  if (!user) {
-    redirect("/");
-  }
+  // if (!user) {
+  //   redirect("/");
+  // }
+
   return (
     <div>
       <header className="sticky top-0 z-40 w-full border-b bg-background">
@@ -44,7 +45,7 @@ export async function SiteHeader() {
                 </div>
               </Link>
               <ThemeToggle />
-              <UserNav user={user} />
+              {user && <UserNav user={user} />}
             </nav>
           </div>
         </div>
@@ -53,7 +54,9 @@ export async function SiteHeader() {
         <div className="container flex h-16 items-center space-x-4 sm:justify-between sm:space-x-0">
           <div className="flex flex-1 items-center justify-end space-x-4">
             <nav className="flex items-center space-x-1">
-              <SecondaryNav user={user} items={siteConfig.secondaryNav} />
+              {user && (
+                <SecondaryNav user={user} items={siteConfig.secondaryNav} />
+              )}
             </nav>
           </div>
         </div>
